Moved ControlTypes to Models
