+++ app/src/main/java/ca/coffeeshopstudio/gaminginterfaceclient/models/ControlTypes.java
@@ -1,4 +1,4 @@
-package ca.coffeeshopstudio.gaminginterfaceclient.views;
+package ca.coffeeshopstudio.gaminginterfaceclient.models;
 
 import android.content.Context;
